Completed Development Version of Send Grid User Sign Up Email Verification

+++ timer-project-nest-js-backend/src/timers/timers.service.ts
@@ -5,6 +5,7 @@ import { Timer } from './entities/timer.entity';
 import { UsersService } from '../users/users.service';
 import { CreateTimerDTO } from './dto/create-timer.dto';
 import { User } from '../users/entities/user.entity';
+import { BadRequestException } from '../exception/bad-request.exception';
 import { TimersUtility } from './timers.utility';
 import { NotFoundException } from '../exception/not-found.exception';
 import { GuestTimer } from './objects/guest-timer';
@@ -21,11 +22,15 @@ export class TimersService {
 
   async createTimer(email: string, createTimerDTO: CreateTimerDTO): Promise<Timer> {
     const user: User = await this.usersService.retreiveUser(email);
+    if(user.numberOfTimers === 3) {
+      throw new BadRequestException('Cannot create more than 3 timers.');
+    }
     const timer: Timer = this.timersRepository.create({ ...createTimerDTO, user: user });
     const savedTimer: Timer = await this.timersRepository.save(timer);
     const completedTimerStats: { end: Date, numberOfBreaks: number } = await this.timerUtility.completeCreateTimer(savedTimer);
     await this.timersRepository.update({ id: savedTimer.id }, { endTime: completedTimerStats.end, numberOfBreaks: completedTimerStats.numberOfBreaks });
     const completeTimer: Timer = await this.retreiveTimer(savedTimer.id);
+    this.usersService.updateUserTimerCount(user.email, true, user.numberOfTimers);
     return completeTimer;
   }
 
@@ -89,8 +94,13 @@ export class TimersService {
     return guestTimer;
   }
 
-  async removeTimer(id: string):Promise<void> {
+  async removeTimer(email: string, id: string) {
+    const user: User = await this.usersService.retreiveUser(email);
+    if(user.numberOfTimers === 0) {
+      throw new BadRequestException('No timers to delete.');
+    }
     const timer: Timer = await this.retreiveTimer(id);
     await this.timersRepository.remove(timer);
+    this.usersService.updateUserTimerCount(user.email, false, user.numberOfTimers);
   }
 }
